Change way of running VTS in tests
This model of getting the war via maven
is copied from omatsivut .
This might work more reliably than the
previous solution.

It is a bit of a kludge to refer to the
runner via reflection, but nevertheless
I find it cleaner to have the VTS runner
in src/test instead of src/main.